// src/enums/crm/basicProduct.ts
export enum ECreateBasicProductParams {
  CATEGORY = 'category',
  COST = 'cost',
  COST_CURRENCY = 'costCurrency',
  DESCRIPTION = 'description',
  LEATHER_ARTICLE = 'leather',
  PUNCH_PITCH = 'punchPitch',
  SIZE = 'size',
  TITLE = 'title',
}
export enum EUpdateBasicProductParams {
  ID = '_id',
  ASSIGNMENTS = 'assignments',
  CATEGORY = 'category',
  COST = 'cost',
  COST_CURRENCY = 'costCurrency',
  DESCRIPTION = 'description',
  LEATHER = 'leather',
  PHOTOS = 'photos',
  PUNCH_PITCH = 'punchPitch',
  SIZE = 'size',
  TITLE = 'title',
}